Import the NOAA 2004 Titanic ROV data which includes 3D coordinates and timestamps of events of interesting objects in the debris field, removed all the extra data since it was 581mb
https://www.ncei.noaa.gov/maps/oer-digital-atlas/mapsOE.htm?lat=1&lng=125&zoom=7&survey=Titanic%252520Expedition%2525202004 and search for “Titanic”. It shows two missions from 2003 and 2004, and it includes extra metadata including the lat-long coordinates of each submarine, as well as a transcript of what the submarine was looking at.

2004 complete archive which is 581mb total uncompressed: https://www.ncei.noaa.gov/archive/archive-management-system/OAS/bin/prd/jquery/download/72311.1.1.tar.gz

Hercules is the NOAA ROV, and Argus is a support vehicle that floats above holding the fiber tether.
I removed all the surface ship and Argus data, as well as the raw sensor data for the water conditions, and just kept the 3D track and event log which is the most useful for now.